requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
azure-communication-callautomation==1.3.0
azure-cognitiveservices-speech==1.37.0
openai==1.35.3
pydantic-settings==2.3.4
python-dotenv==1.0.1
